+++ docs/release-notes.rst
@@ -21,6 +21,7 @@ New API
 - Support of ``resource.postData`` in callback ``webpage.onResourceRequested``.
 - Support of `navigator.mozTCPSocket <https://developer.mozilla.org/en-US/docs/WebAPI/TCP_Socket>`_ is enabled
   for the main script. It needs at least XulRunner/Firefox 25 to use this API.
+- New property ``slimer.geckoVersion``
 
 Fixed bugs
 ----------