+++ docs/release-notes.rst
@@ -23,6 +23,8 @@ New API
   for the main script. It needs at least XulRunner/Firefox 25 to use this API.
 - New property ``slimer.geckoVersion``
 - Compatible with Firefox/XulRunner 25b4
+- Key shortcut ctrl+w on windows to quit SlimerJS (main window) or just close the window.
+  (useful when the scripts ends without exit)
 
 Fixed bugs
 ----------