+++ docs/release-notes.rst
@@ -5,6 +5,19 @@
 Release Notes
 =============
 
+version 0.8.4
+=============
+
+Not released yet.
+
+
+Fixed bugs
+----------
+
+- redirection were not well supported when the requested URL is changed by the server
+  by adding a trailing slash (issue #90)
+
+
 version 0.8.3
 =============